Split entrance stage entities into 2 files
