+++ config/splat.us.stnp3.yaml
@@ -44,6 +44,7 @@ segments:
       - [0xD98, .data, st_debug]
       - [0xE18, .data, e_breakable]
       - [0xE78, .data, entrance_stage_entities]
+      - [0x1008, .data, entrance_stage_entities_2]
       - [0x122C, data]
       - [0x1300, data]
       - [0x14B4, .data, gaibon]
@@ -132,7 +133,7 @@ segments:
       - [0x168F4, .data, tile_data] # tile data
       - [0x21CF4, .data, tile_data] # tile definitions
       - [0x2A124, .data, gen/sprites]
-      - [0x31EA0, .rodata, entrance_stage_entities] # EntityStairwayPiece
+      - [0x31EA0, .rodata, entrance_stage_entities_2] # EntityStairwayPiece
       - [0x31EBC, .rodata, e_explosion_puff_opaque] # EntityExplosionPuffOpaque
       - [0x31ED0, .rodata, slogra] # EntitySlogra .rodata, 36990
       - [0x31EDC, .rodata, gaibon] # EntityGaibon .rodata, 36990
@@ -160,6 +161,7 @@ segments:
       - [0x3246C, c, st_debug]
       - [0x326FC, c, e_breakable]
       - [0x32830, c, entrance_stage_entities]
+      - [0x32F30, c, entrance_stage_entities_2]
       - [0x363FC, c, e_explosion_puff_opaque]
       - [0x36990, c, slogra]
       - [0x378BC, c, gaibon]
