+++ src/st/np3/entrance_stage_entities_2.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "np3.h"
+
+#include "../entrance_stage_entities_2.h"
